feat: next-auth set
