+++ app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,18 @@
+import NextAuth from 'next-auth';
+import GoogleProvider from 'next-auth/providers/google';
+
+const authOptions = {
+  // Configure one or more authentication providers
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  session: {
+    maxAge: 60 * 60 * 24,
+  },
+};
+
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
