+++ app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from 'next';
 import { dm_sans, noto_sans } from './fonts';
 import './globals.css';
+import Provider from './provider';
 
 export const metadata: Metadata = {
   title: '이너-프',
@@ -18,7 +19,7 @@ export default function RootLayout({
         lang="en"
         className={`${noto_sans.variable} ${dm_sans.variable} font-dm-sans`}
       >
-        {children}
+        <Provider>{children}</Provider>
       </body>
     </html>
   );
