+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "next": "15.1.7",
+    "next-auth": "^4.24.11",
     "next-themes": "^0.4.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
