+++ src/features/socialButton/ui/BaseButton.tsx
@@ -1,5 +1,4 @@
 'use client';
-
 import { twAllMerge } from '@/shared/utils';
 
 const BaseButton = ({
