+++ src/features/socialButton/ui/Kakao.tsx
@@ -1,5 +1,4 @@
 'use client';
-
 import BaseButton from './BaseButton';
 
 const InternalKaKao = ({
