+++ src/features/socialButton/ui/Naver.tsx
@@ -1,5 +1,4 @@
 'use client';
-
 import BaseButton from './BaseButton';
 
 const InternalNaver = ({
